Add Freshservice integration
Freshservice is a help-desk product in the FreshWorks family, alongside
Freshdesk and Freshrelease.  This commit adds the start/stop timer
button to the ticket detail page.

On Chrome, Freshservice's CSS overrides the left margin of span
elements, resulting in incorrect padding between the button icon and
text.  Adding the !important flag to button span's CSS corrects this.