Use GitLab reference identifiers in GitLab integration
Update the task name for Gitlab issues and MRs to be prefixed with the full GitLab reference (including group and project identifiers) instead of just the id number.

This matches the text used for cross-project references in Markdown on GitLab. See: https://docs.gitlab.com/ee/user/markdown.html#gitlab-specific-references