Initial implementation of project view, support both Issues and Drafts.
Warning:  Requires at least one task/card to be linked to an issue, to procure the Repository name to feed into the Project value of the new clockify task.

I do not see any other way to deduce it from the raw page content, nor localStorage of GitHub.com. Nor the url.
Fallback is to use the Name of the Project.

Note: This is not for the slideout, only the main card display of the Board.
Also need to work on Table view too.