more attempts at making this work.
For unknown reasons, the hidden internal dom property __reactProps will never exist unless chrome dev tools is open and you navigate to the DOM element, exposing the properties in the properties view.

I do not know why, and this was already a major hack.
I think it's time to look for a different approach.
Committing code which I will soon delete, for posterity.