I've settled on the idea that we likely will not be able to determine Project name.
So we'll fallback on the name of the Project.
I've extrapolated the common logic, to try and scan against the existing known projects of clockify, using a substring match. And fallback further against any detected available issue link.
It's the best we can do.

Also implemented table handler in full, testing good.