Attempting to fix for mutator when user switches tabs (code was reliant upon manual page refresh).
when changing tabs, React is render cycling the entire document tree.

However it would appear that two issues are present:
a. webpack is transpiling the Class definition for our helper method into native, and it's already declared producing an error preventing execution.
b. import() statements are not permitted because the transpiled "integration" is not an esm module.

So it looks like I will have to rewrite my class helper into a const func, which degrades the async initializer/constructor approach.
Ugh.