Feature/integration GitHub project perspective + drafts - 2024 09 (#275)
* Initial implementation of project view, support both Issues and Drafts.

Warning:  Requires at least one task/card to be linked to an issue, to procure the Repository name to feed into the Project value of the new clockify task.

I do not see any other way to deduce it from the raw page content, nor localStorage of GitHub.com. Nor the url.
Fallback is to use the Name of the Project.

Note: This is not for the slideout, only the main card display of the Board.
Also need to work on Table view too.

* attempting to access react internals... since there is no other way to access data within github's ui via html.

* more attempts at making this work.

For unknown reasons, the hidden internal dom property __reactProps will never exist unless chrome dev tools is open and you navigate to the DOM element, exposing the properties in the properties view.

I do not know why, and this was already a major hack.
I think it's time to look for a different approach.
Committing code which I will soon delete, for posterity.

* I've settled on the idea that we likely will not be able to determine Project name.

So we'll fallback on the name of the Project.
I've extrapolated the common logic, to try and scan against the existing known projects of clockify, using a substring match. And fallback further against any detected available issue link.
It's the best we can do.

Also implemented table handler in full, testing good.

* mutator cleanup, old approach. Meant to include this in previous commit.

* Attempting to fix for mutator when user switches tabs (code was reliant upon manual page refresh).

when changing tabs, React is render cycling the entire document tree.

However it would appear that two issues are present:
a. webpack is transpiling the Class definition for our helper method into native, and it's already declared producing an error preventing execution.
b. import() statements are not permitted because the transpiled "integration" is not an esm module.

So it looks like I will have to rewrite my class helper into a const func, which degrades the async initializer/constructor approach.
Ugh.

* rewrote it as a function. const was unsatisfactory and still produced the error. I think webpack esm -> native , s happy now.

Tested good. Seeing fresh results whenever user switches tabs on github.

* removing uneccessay log bloat.

* Fix:  tasks which do not have an issue were not being targeted on the table selector.

Was caused by comparison of zero-based index with a value of 0 to equivalency of false.

* looks like sidepanel is already implemented. We just need to tweak it to support non-issue tasks.

moving code order of ops, so we can reuse my helper class.
And next commit will contain the actual fix.

* Fix: non-issue slideout panel tasks

* null exception handler, for Issue List, where we expect to fail on this routine.

* Bug: typo in code, wasn't processing page url (default of plugin for Issue View).